+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.6.11
+
+- Also apply the accessoryType fix on iOS 12
+
+## 1.6.10
+
+- Fix cut view when UITableViewCell's accessoryType is set to disclosureIndicator in a storyboard.
+
 ## 1.6.9
 
 - iOS 13 compatibility fixes.